pyproject.toml
[tool.poetry]
name = "agent-ai"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
openai = "^1.51.0"
uvicorn = "^0.31.0"
twilio = "^9.3.3"
python-dotenv = "^1.0.1"
python-docx = "^1.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
